use map_solver with preprocessing
This will allow to extract map goals into standalone goals if we want
to try other tools, and it makes the map_solver run faster:

Originally, FlattenExprOld.v took 326s to compile.
If we replace map_solver with "admit" (or "case TODO", to be precise),
it takes 119s to compile, so the map_solver without preprocessing
took 207s.
Now, FlattenExprOld.v takes 216s, so the map_solver with preprocessing
takes 97s.
So the preprocessing is a 2x speed improvement for the map_solver,
and a 1.5x speed improvement for the whole file.