lia is significantly slower than omega
The sum of all invocation times for lia is 5x higher than for omega.
If we ignore the 3 highest values of each (out of 90000 values),
lia is 64x slower than omega.
And if we add 1ms for each invocation which was reported as 0ms to
account for lack of fine granularity, lia is still 16x slower
(again with 3 highest values of each ignored).