simplest possible riscv_ext_spec instance which accepts any MMIO
So at the moment we still prove the same as before, but at least it
is parametrized.
To prove stronger guarantees about what MMIO occurs, one could prove a
postcondition about the final trace (but then we do not learn anything
about traces of programs not yet run to completion), or, thanks to this
new parametrization, add more constraints to riscv_ext_spec which would
the RiscV machine make fail if they are violated.