Minimal compatibility so that NotationsCustomEntry.v compiles.
In the line:

  Notation "x = ( e )" := (set x e) (in custom bedrock_cmd at level 0, x constr at level 0, e custom bedrock_expr).

The reference "x constr at level 0" was wrongly interpreted as "x
custom bedrock_expr". So that NotationsCustomEntry directly compiles,
I simulated the old (buggy) behavior:

  Notation "x = ( e )" := (set x e) (in custom bedrock_cmd at level 0, x custom bedrock_cmd, e custom bedrock_expr).

I tried to respect the suspected intent and to instead remove the
"constr:(_)" in line 183 (and following) but the grammar does not
factorize correctly the call to "global" and to "constr at level 0",
leading to a failure at line 185, so I renounced to investigate this
direction further.