lib/compiler: Try default values for evars eagerly
A lemma may produce a goal [ |- g ?ev] that leads to ?ev being instantiated; but
in other cases, the goal might not be solvable unless ?ev is instantiated first.
So, as a last resort, if we have another goal [ |- DefaultValue _ ?ev], we
instantiate ?ev with its default value and try to solve again.