lib/loops: Rebuild invariant inference using on the new algorithm
Instead of looking for values, this version looks for names.  The previous
version generalized over the `from` and `init` parts of the loop; this version
is a lot more predictable: it generalizes the variables that the results of the
loop are stored in (or the objects these variables point to).