in RunInstruction/load_save_regs lemmas, introduce extra Exec
var to capture all instruction memory.
This ensures that lemmas that modify the memory do not
modify the symbolic representation of the instruction
memory (before, they changed the address of the code
being executed from an offset calculated from the code
to the symbolic value of the pc).
It also ensures that lemmas that do not modify the
memory at all do not require that executable memory is
unfolded all the way to exposing the instruction being
run.

Also apply this trick to GhostConsts, so that the same
advantages apply when invoking the IH.

Also replacing simp by fwd.

[wip, does not compile]