replace at_addr by sepcl, which also takes a value,
so that use sites of sepcl can remain unchanged if we want
to swap addr/value order of all sep predicates, and we
don't need two versions of each predicate any more
(deleting SepAddrLast.v).
Moreover, sepcl enables notations that want to identify
not only "the address", but also "the value" and "the
predicate" of a clause.