expose the whole goal to the split_sepclause search tactic
so that it can context-match on a `length` precondition to
get the length desired by the callee, so we don't need
`with_len` any more.

Also switch from `seps [...; emp (continuation)]` to
`seps [...] /\ continutation`, adapting all cancellation
lemmas to a goal of shape `seps [...] /\ continuation`
instead of `seps [...]`, and then use continutation:=True
if there is none.
Adapting the cancellation infrastructure to include
`_ /\ continuation` everywhere is a bit of a disadvantage,
but has the advantage that split_sepclause search hints
can identify the continuation (which might be of shape
`more_sideconditions /\ actual_continutation`) without
having to iterate to the last element of the `seps`,
so the context match on `more_sideconditions` to find
a `length ?vs = desired_length` can focus more easily
on just the part where we expect to find the length
equality.

Moreover, in LiveVerif, use get_option wrapped around
Memory.load/Memory.store as preconditions of wp_load/store
instead of inlined separation logic preconditions, so
that automation can check more easily whether we're solving
a load sidecondition or a store sidecondition, because the
former requires clearing the split_merge_lemma stack without
using it, whereas the latter requires applying all the merges.