autorew: home-made autorewrite, extensible through typeclass search,
can also rewrite with ring_simplify and with local hypotheses.

In swap.v, use autorew instead of rewrite_db (and instead of
word_simpl_step_in_hyps)

Compilation time of swap.v remains about the same
(from 69.47 down to 68.81), and Ltac Profile shows
that most of the time is spent in ZnWords sideconditions.