0-initialize stackalloc arrays in ToCString
Bedrock2 semantics leave the contents of stack-allocated arrays
unspecified, but still model memory as a map from words to bytes.
This implies that the freshly-allocated bytes must read as some fixed
value, as opposed to ISO where reading uninitialized variables is
undefined behavior.

bedrock2Examples.stackalloc.stacknondet was supposed to demonstrate the
friendlier semantics of bedrock2, but it wasn't finished until now.
This commit adds a proof example that reading the same uninitialized
addresses twice in Bedrock2 returns the same value, and adds this
example as a test.

It was manually confirmed using that valgrind rejects the test before
the ToCString fix but accepts the updated version.