fix ZnWords usage in lightbulb.v
In `replace t with u in * by tac`, if the pattern `t` contains holes
that can't be inferred because `t` does not appear anywhere in the
goal, but running `tac` instantiates them, `replace` will run `tac`
to produce an unused proof and succeed. But if `tac` fails (which
is the case in the new ZnWords that rejects goals with evars),
`replace` will fail as well.