basic support for loading/storing words (instead of only Zs)
convention:

| type  | alias    | predicate        | access in source |
|-------|----------|------------------|------------------|
| Z     | uint_t N | uint N val addr  | loadN/storeN     |
| word  | word     | uintptr val addr | load/store       |

where N denotes the number of bits and can be 8, 16 or 32.
64-bit values currently can only be accessed as words, not as Zs,
or actually, not even as words, because all of LiveVerif is 32-bit
for now.
The 'alias' column gives the type to use to make sure that the
create_predicate tactic knows how much space to use for a Z.