Add wf, start writing elaborate_wf
Add a "well-formedness" judgement for expressions which can be used to
state that `elaborate` correctly handles scope (i.e., an elaborate
expression never has free variables). The soundness of the types
assigned by `elaborate` is already enforced via the GADT.

Some parts of `elaborate_wf` are still unfinished; specifically, the
parts that require proving that two type variables are equal and the
parts relating to records.