"same start address and same size" does not imply "safe to cancel"
because the clause to be canceled could be a bst node with different
subtrees than the candidate clause.
Stronger condition: both clauses need to be contiguous (which implies
that they have no pointers to other memory that might differ) OR
cancelling the clause with the candidate clause works right away
(without splitting it into steps and potentially leaving some subgoals
open for manual proving, which requires knowing that the cancellation
decision is safe).