Deshadow variable names in elaborate
Making deshadowing a prerequisite for wf means that Compile.v will not
need to deal with the possibility of variable shadowing. The
is_square_exists proof in SamplePrograms.v is left admitted for now and
will need to be updated to account for the fact that variable names in
general no longer match between the pexpr language and the expr
language.