split_range_from_hyp: don't simplify index and offset
index and offset only appear in the intermediate symbolic states created
by the split_off_... lemmas, but not in the final remerged user-visible
states, and even if they did, they'd get simplified at the end by
after_command_simpl_hook, so it's ok if we don't simplify them, from the
user's perspective.
From the proving machinery's perspective, term sizes might become a bit
bigger, but probably ok because of savings from not simplifying, and
in find_field_index_from_offset, we now need xlia instead of reflexivity.
The main advantage is that more goals become provable: If we
access a record field at address `p = a ^+ /[8]`, the offset becomes
`p ^- a`, and it's unclear how to tell `bottom_up_simpl_in_goal` that
here it's ok and required to `subst p` before simplifying, whereas xlia
knows how to deal with equalities in hypotheses.