get started with fixed size malloc,
which has a different out-of-memory predicate than one_size_malloc:
the allocator predicate just takes an argument n_remaining that
indicates how many blocks we still can allocate, and usually, clients
won't keep track of this number precisely, but in order to be allowed
to fail, they will need to give a low (eg 0 or 1) n_remaining in
their postcondition.
This should fix the cases like critbit_insert where the second of
two allocations fails: Now we can free the first, and return an allocator
with n_remaining=1.