Parametrize FlatImp.exec by compiler phase
The spilling phase adds some boilerplate inside functions and around
function calls, so we need to account for it by making things
artificially more expensive before the spilling phase.

In an effort to avoid monolithic commits, this is a partial commit, so
the code definitely doesn't compile in the current state since lots of
things changed as a side effect.

There are a few comments that say "cost_something constraint:", which
document the various constraints on the magic numbers chosen as upper
bounds for the costs (100, 95, 50). These should probably change to
better numbers at some point.