get stuck in sequence (and while) case.  IH is no good!
(forall k, pick_spH (k ++ kH) = pick_spL (rev kL ++ snd (fst (dtransform_stmt_trace eH (rev k, sH, used_after)))))

^this part is too strong.  not preserved when passing to first subcommand of
seq.  we don't want equality for all k; we just want equality for short
enough k.  but this would be obnoxious to work with, I think.

if I had the continuation here, then putting that extra degree of freedom
in there allows it to be true for all k!  I think putting the continuation
back would be much less pain, in the end.  the short-circuiting thing in
writing the function is really annoying, anyway

There are a couple dimensions along which these trace-transformation
functions proofs can vary: we can either have the deterministic stackalloc
or not, and I we can either have the function written in CPS or not.  Before
this, I had explored three points in this space, and they all worked out
fine.  Here I explored the fourth, deterministic stackalloc + not-CPS, and
it did not work as well.