straightline: do not match below binders
When flattening seps, straightline was matching below binders which it
couldn't simplify. This caused some slowdown because the match and
simplify happened on each run, each hypothesis, and each wrongly
bracketed sep in that hypothesis.

Simply replacing _ with ?_a resolves it because the matching can't bind
something that's unresolved to _a.

Also adding some tests for ProgramLogic, for this issue and the previous
one I fixed where straightline was destructing context variables, which
it shouldn't.